# dashboard/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Industrial Time-Series Monitoring API",
    "description": "\nREST API over a historical archive of condition-monitoring data from a metro train's\ncompressor Air Production Unit (UCI MetroPT-3, DOI 10.24432/C5VW3R).\n\n**What the data is.** 22,754,220 readings - 1,516,948 scans of 15 sensors, February to\nSeptember 2020. Coverage is 82.4%: the remaining 17.6% of the timeline has no data, in\n331 recorded gaps.\n\n**Three things this API will not do to you.**\n\n* *Silently truncate.* Any series that hits a result cap is returned with\n  `truncated: true`.\n* *Hide data quality.* 3.35% of the archive is held data from a logger freeze - non-null,\n  in range, and not a measurement. It is returned with its quality code, never filtered\n  out behind your back.\n* *Return an unusable number of points.* `resolution=auto` picks raw, hourly or daily\n  from the width of the window.\n\n**Thresholds are data, not code.** Every severity boundary traces to a row in\n`analytics.SensorBaseline` or to a setpoint published by the equipment owner.\n",
    "contact": {
      "name": "Project documentation",
      "url": "https://archive.ics.uci.edu/dataset/791/metropt+3+dataset"
    },
    "license": {
      "name": "Dataset: UCI MetroPT-3, DOI 10.24432/C5VW3R"
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/readings/{sensor_id}": {
      "get": {
        "tags": [
          "readings"
        ],
        "summary": "Historical readings for one sensor",
        "description": "Return a series, choosing a resolution that a client can actually draw.\n\nThe response always states the resolution it served and whether a cap was hit, so a\ntruncated series can never be mistaken for a complete one.",
        "operationId": "get_readings_readings__sensor_id__get",
        "parameters": [
          {
            "name": "sensor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Numeric id or sensor code, e.g. 15 or TP3.",
              "title": "Sensor Id"
            },
            "description": "Numeric id or sensor code, e.g. 15 or TP3."
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive start of the window (ISO 8601). Defaults to the end of the archive minus the default window.",
              "title": "Start"
            },
            "description": "Inclusive start of the window (ISO 8601). Defaults to the end of the archive minus the default window."
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exclusive end of the window (ISO 8601).",
              "title": "End"
            },
            "description": "Exclusive end of the window (ISO 8601)."
          },
          {
            "name": "resolution",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Resolution",
              "description": "'auto' picks raw / hourly / daily from the width of the window. Ask for a specific resolution only if you need it - six months of raw data is 1.8 million points.",
              "default": "auto"
            },
            "description": "'auto' picks raw / hourly / daily from the width of the window. Ask for a specific resolution only if you need it - six months of raw data is 1.8 million points."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadingSeries"
                }
              }
            }
          },
          "404": {
            "description": "No such sensor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Invalid time range",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/trends/{sensor_id}": {
      "get": {
        "tags": [
          "readings"
        ],
        "summary": "Trend with rolling statistics",
        "description": "Hourly series with a rolling mean, rolling spread and rate of change.\n\nRate of change is per hour of **elapsed** time rather than per bucket, because 17.6%\nof the archive is missing: treating a 48-hour gap as one step would report a violent\nswing where there was simply no data.",
        "operationId": "get_trend_trends__sensor_id__get",
        "parameters": [
          {
            "name": "sensor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Numeric id or sensor code.",
              "title": "Sensor Id"
            },
            "description": "Numeric id or sensor code."
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive start of the window (ISO 8601). Defaults to the end of the archive minus the default window.",
              "title": "Start"
            },
            "description": "Inclusive start of the window (ISO 8601). Defaults to the end of the archive minus the default window."
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exclusive end of the window (ISO 8601).",
              "title": "End"
            },
            "description": "Exclusive end of the window (ISO 8601)."
          },
          {
            "name": "operating_state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/OperatingState"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Restrict the trend to one operating state. Strongly recommended for anything statistical: the compressor is off 54.65% of the time, so a state-blind mean blends idle and running behaviour.",
              "title": "Operating State"
            },
            "description": "Restrict the trend to one operating state. Strongly recommended for anything statistical: the compressor is off 54.65% of the time, so a state-blind mean blends idle and running behaviour."
          },
          {
            "name": "window_hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 168,
              "minimum": 2,
              "description": "Rolling window, in hourly buckets.",
              "default": 24,
              "title": "Window Hours"
            },
            "description": "Rolling window, in hourly buckets."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrendSeries"
                }
              }
            }
          },
          "404": {
            "description": "No such sensor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Invalid time range",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/readings": {
      "get": {
        "tags": [
          "readings"
        ],
        "summary": "Readings for several sensors at once",
        "description": "Fetch several series in one round trip, for a multi-tag chart.\n\nCapped at eight sensors. The point cap is per series, so an uncapped list would let a\nsingle request ask for 15x the ceiling.",
        "operationId": "get_multi_readings_readings_get",
        "parameters": [
          {
            "name": "sensors",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Comma-separated sensor ids or codes, e.g. 'TP2,TP3,OIL_TEMPERATURE'.",
              "examples": [
                "TP2,TP3"
              ],
              "title": "Sensors"
            },
            "description": "Comma-separated sensor ids or codes, e.g. 'TP2,TP3,OIL_TEMPERATURE'."
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive start of the window (ISO 8601). Defaults to the end of the archive minus the default window.",
              "title": "Start"
            },
            "description": "Inclusive start of the window (ISO 8601). Defaults to the end of the archive minus the default window."
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exclusive end of the window (ISO 8601).",
              "title": "End"
            },
            "description": "Exclusive end of the window (ISO 8601)."
          },
          {
            "name": "resolution",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Resolution",
              "default": "auto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadingSeries"
                  },
                  "title": "Response Get Multi Readings Readings Get"
                }
              }
            }
          },
          "404": {
            "description": "No such sensor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Invalid time range",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AggregatePoint": {
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "value": {
            "type": "number",
            "title": "Value",
            "description": "Mean over the bucket."
          },
          "min_value": {
            "type": "number",
            "title": "Min Value"
          },
          "max_value": {
            "type": "number",
            "title": "Max Value"
          },
          "sample_count": {
            "type": "integer",
            "title": "Sample Count"
          },
          "coverage_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage Pct",
            "description": "Samples present as a percentage of a fully-covered bucket at the measured 10 s interval. The archive is 82.4% covered overall, so many buckets hold fewer samples than a naive reader would assume."
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "value",
          "min_value",
          "max_value",
          "sample_count"
        ],
        "title": "AggregatePoint",
        "description": "One bucket of an aggregated series."
      },
...
      "OperatingState": {
        "type": "string",
        "enum": [
          "OFF",
          "OFFLOADED",
          "LOADED",
          "STARTING"
        ],
        "title": "OperatingState"
      },
      "ProblemDetail": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "title": "ProblemDetail",
        "description": "RFC 7807 problem details.\n\nA machine-readable error shape, so a client can branch on ``type`` rather than\npattern-matching prose that will be reworded eventually.",
        "example": {
          "detail": "start (2020-06-01T00:00:00) must be earlier than end (2020-05-01T00:00:00).",
          "instance": "/readings/14",
          "status": 422,
          "title": "Invalid time range",
          "type": "/problems/invalid-time-range"
        }
      },
      "ReadingPoint": {
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "quality": {
            "type": "string",
            "title": "Quality"
          },
          "quality_usable": {
            "type": "boolean",
            "title": "Quality Usable"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "value",
          "quality",
          "quality_usable"
        ],
        "title": "ReadingPoint",
        "description": "One point of a raw series."
      },
      "ReadingSeries": {
        "properties": {
          "sensor_id": {
            "type": "integer",
            "title": "Sensor Id"
          },
          "sensor_code": {
            "type": "string",
            "title": "Sensor Code"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "resolution": {
            "$ref": "#/components/schemas/Resolution",
            "description": "The resolution actually served, which may differ from the one requested when 'auto' was used or when a cap was applied."
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "title": "End"
          },
          "point_count": {
            "type": "integer",
            "title": "Point Count"
          },
          "truncated": {
            "type": "boolean",
            "title": "Truncated",
            "description": "True when a result cap was reached and the series is incomplete. Never silently trimmed.",
            "default": false
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/ReadingPoint"
            },
            "type": "array",
            "title": "Points"
          },
          "aggregates": {
            "items": {
              "$ref": "#/components/schemas/AggregatePoint"
            },
            "type": "array",
            "title": "Aggregates"
          }
        },
        "type": "object",
        "required": [
          "sensor_id",
          "sensor_code",
          "unit",
          "resolution",
          "start",
          "end",
          "point_count"
        ],
        "title": "ReadingSeries"
      },
      "Resolution": {
        "type": "string",
        "enum": [
          "auto",
          "raw",
          "hourly",
          "daily"
        ],
        "title": "Resolution",
        "description": "Time resolution of a series.\n\n``AUTO`` lets the server choose based on the requested window, which is what stops a\nsix-month request from trying to return 1.8 million points per sensor."
      },
...
      "TrendPoint": {
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "rolling_mean": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rolling Mean"
          },
          "rolling_std": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rolling Std"
          },
          "rate_of_change_per_hour": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate Of Change Per Hour"
          },
          "sample_count": {
            "type": "integer",
            "title": "Sample Count"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "value",
          "sample_count"
        ],
        "title": "TrendPoint"
      },
      "TrendSeries": {
        "properties": {
          "sensor_id": {
            "type": "integer",
            "title": "Sensor Id"
          },
          "sensor_code": {
            "type": "string",
            "title": "Sensor Code"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "operating_state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OperatingState"
              },
              {
                "type": "null"
              }
            ],
            "description": "When set, the trend covers only periods in this operating state. The compressor is off 54.65% of the time, so a state-blind trend blends idle and running behaviour into a number describing neither."
          },
          "window_hours": {
            "type": "integer",
            "title": "Window Hours"
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "title": "End"
          },
          "point_count": {
            "type": "integer",
            "title": "Point Count"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/TrendPoint"
            },
            "type": "array",
            "title": "Points"
          },
          "summary": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "sensor_id",
          "sensor_code",
          "unit",
          "window_hours",
          "start",
          "end",
          "point_count",
          "points"
        ],
        "title": "TrendSeries"
      },
...
    }
  },
...
}